Change how JS get the template element
Because the document object doesn't include what's inside combobox.html
after being imported, I have to change the scope to use
document.currentScript.ownerDocument for the combobox's prototype
script.
For that reason, I reformatted the combobox object slightly to include
the template as a property to the object itself for ease of references.