Test ul shadow elem is hidden by default
Create a module for testing things that should happen on element
creation. Define setup and teardown for the module.
Have to switch to inject the element to the page instead of using
document.createElement() because apparently that doesn't get the style
applied to. There's no way to test the style of the element. The
injected element is removed as part of the tear down.