Update the style of the input field and button
Change the button to span to allow user to tab directly out of the
element to another element, instead of from input to button element.
Replace the arrow-down character with another that looks like it has
better padding-size better top and bottom.
Move the arrow character inside the input field and add padding to the
input field to prevent typing over the arrow character.